# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Agent Orchestrator HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/mobile/disable:
    post:
      operationId: disableMobile
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileStatusResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Disable the Connect Mobile LAN bridge
      tags:
      - mobile
  /api/v1/mobile/enable:
    post:
      operationId: enableMobile
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileStatusResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Enable the Connect Mobile LAN bridge and issue a fresh password
      tags:
      - mobile
  /api/v1/mobile/regenerate:
    post:
      operationId: regenerateMobile
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileStatusResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Rotate the Connect Mobile password, dropping any connected phone
      tags:
      - mobile
  /api/v1/mobile/status:
    get:
      operationId: getMobileStatus
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileStatusResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Forbidden
      summary: Check whether Connect Mobile's LAN bridge is enabled
      tags:
      - mobile
...
  /api/v1/sessions:
    get:
      operationId: listSessions
      parameters:
      - description: Project id filter.
        in: query
        name: project
        schema:
          description: Project id filter.
          type: string
      - description: When true, return non-terminated sessions; when false, return
          terminated sessions.
        in: query
        name: active
        schema:
          description: When true, return non-terminated sessions; when false, return
            terminated sessions.
          type:
          - "null"
          - boolean
      - description: When true, return only orchestrator sessions.
        in: query
        name: orchestratorOnly
        schema:
          description: When true, return only orchestrator sessions.
          type:
          - "null"
          - boolean
      - description: When true, return only fresh non-terminated sessions.
        in: query
        name: fresh
        schema:
          description: When true, return only fresh non-terminated sessions.
          type:
          - "null"
          - boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSessionsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: List sessions
      tags:
      - sessions
    post:
      operationId: spawnSession
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpawnSessionRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpawnSessionResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Spawn a new agent session
      tags:
      - sessions
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    AgentConfig:
      properties:
        model:
          type: string
        permissions:
          type: string
      type: object
...
    ControllersSessionView:
      properties:
        activity:
          $ref: '#/components/schemas/DomainActivity'
        agentConfig:
          $ref: '#/components/schemas/AgentConfig'
        branch:
          type: string
        createdAt:
          format: date-time
          type: string
        displayName:
          type: string
        harness:
          type: string
        id:
          type: string
        isTerminated:
          type: boolean
        issueId:
          type: string
        kind:
          type: string
        previewRevision:
          format: int64
          type: integer
        previewUrl:
          type: string
        projectId:
          type: string
        prompt:
          type: string
        prs:
          items:
            $ref: '#/components/schemas/SessionPRFacts'
          type: array
        scmStatus:
          enum:
          - pr_open
          - draft
          - ci_failed
          - review_pending
          - changes_requested
          - approved
          - mergeable
          - merged
          type: string
        status:
          enum:
          - working
          - pr_open
          - draft
          - ci_failed
          - review_pending
          - changes_requested
          - approved
          - mergeable
          - merged
          - needs_input
          - exited
          - idle
          - terminated
          - no_signal
          type: string
        terminalHandleId:
          type: string
        terminateOnPrMerge:
          type: boolean
        updatedAt:
          format: date-time
          type: string
        workspacePath:
          type: string
      required:
      - id
      - projectId
      - kind
      - activity
      - isTerminated
      - terminateOnPrMerge
      - createdAt
      - updatedAt
      - status
      - prs
      type: object
    ControllersSpawnAttachmentInput:
      properties:
        data:
          type: string
        mimeType:
          type: string
      required:
      - data
      type: object
...
    DomainActivity:
      properties:
        lastActivityAt:
          format: date-time
          type: string
        state:
          type: string
      required:
      - state
      - lastActivityAt
      type: object
...
    ListSessionsResponse:
      properties:
        sessions:
          items:
            $ref: '#/components/schemas/ControllersSessionView'
          type: array
      required:
      - sessions
      type: object
...
    MobileStatusResponse:
      properties:
        enabled:
          type: boolean
        host:
          type: string
        password:
          type: string
        port:
          type: integer
        warning:
          type: string
      required:
      - enabled
      - host
      - port
      - password
      - warning
      type: object
...
    SessionPRFacts:
      properties:
        ci:
          enum:
          - unknown
          - pending
          - passing
          - failing
          type: string
        mergeability:
          enum:
          - unknown
          - mergeable
          - conflicting
          - blocked
          - unstable
          type: string
        number:
          type: integer
        review:
          enum:
          - none
          - approved
          - changes_requested
          - review_required
          type: string
        reviewComments:
          type: boolean
        state:
          enum:
          - draft
          - open
          - merged
          - closed
          type: string
        updatedAt:
          format: date-time
          type: string
        url:
          type: string
      required:
      - url
      - number
      - state
      - ci
      - review
      - mergeability
      - reviewComments
      - updatedAt
      type: object
...
    SpawnSessionRequest:
      properties:
        agentConfig:
          $ref: '#/components/schemas/AgentConfig'
        attachments:
          items:
            $ref: '#/components/schemas/ControllersSpawnAttachmentInput'
          type: array
        branch:
          type: string
        displayName:
          maxLength: 20
          type: string
        harness:
          enum:
          - claude-code
          - codex
          - aider
          - opencode
          - grok
          - droid
          - amp
          - agy
          - crush
          - cursor
          - qwen
          - copilot
          - goose
          - auggie
          - continue
          - devin
          - cline
          - kimi
          - kiro
          - kilocode
          - vibe
          - pi
          - autohand
          - fake
          type: string
        issueId:
          type: string
        kind:
          enum:
          - worker
          - orchestrator
          type: string
        projectId:
          type: string
        prompt:
          maxLength: 4096
          type: string
      required:
      - projectId
      type: object
    SpawnSessionResponse:
      properties:
        promptBytes:
          type: integer
        session:
          $ref: '#/components/schemas/ControllersSessionView'
        systemPromptBytes:
          type: integer
      required:
      - session
      - promptBytes
      - systemPromptBytes
      type: object
...